+++ Makefile.client
@@ -6,7 +6,7 @@
 #    By: oadhesiv <secondfry+school21@gmail.com>    +#+  +:+       +#+         #
 #                                                 +#+#+#+#+#+   +#+            #
 #    Created: 2020/02/29 13:58:56 by oadhesiv          #+#    #+#              #
-#    Updated: 2021/09/05 18:28:55 by oadhesiv         ###   ########.fr        #
+#    Updated: 2021/09/07 16:40:43 by oadhesiv         ###   ########.fr        #
 #                                                                              #
 # **************************************************************************** #
 
@@ -33,7 +33,7 @@ FILES_INIT := $(addprefix init_rtv/, $(FILES_INIT))
 FILES_LOOP :=	loop_hook_common.c loop_sdl.c
 FILES_LOOP := $(addprefix loop/, $(FILES_LOOP))
 
-FILES_IMGUI :=	controls.c imgui_init.c windows.c
+FILES_IMGUI :=	controls.c imgui_init.c windows.c window_controls.c
 FILES_IMGUI := $(addprefix imgui/, $(FILES_IMGUI))
 
 FILES_SDL :=	init_sdl.c sdl_events.c sdl_gl_attributes.c sdl_clear_texture.c
