+++ includes/imgui/window_controls.h
@@ -1,24 +1,20 @@
 /* ************************************************************************** */
 /*                                                                            */
 /*                                                        :::      ::::::::   */
-/*   controls.h                                         :+:      :+:    :+:   */
+/*   window_controls.h                                  :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
 /*   By: oadhesiv <secondfry+school21@gmail.com>    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
-/*   Created: 2021/09/05 14:12:56 by oadhesiv          #+#    #+#             */
-/*   Updated: 2021/09/07 16:40:06 by oadhesiv         ###   ########.fr       */
+/*   Created: 2021/09/07 16:38:18 by oadhesiv          #+#    #+#             */
+/*   Updated: 2021/09/07 16:38:38 by oadhesiv         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-#ifndef CONTROLS_H
-# define CONTROLS_H
+#ifndef WINDOW_CONTROLS_H
+# define WINDOW_CONTROLS_H
 
-# define CIMGUI_DEFINE_ENUMS_AND_STRUCTS
-# include "cimgui.h"
+# include "rtv.h"
 
-void	ig_image_defaults(ImTextureID user_texture_id, const ImVec2 size);
-bool	ig_drag_vector(const char *label, const double *v);
-bool	ig_group_top_default(const char *label);
-bool	ig_group_default(const char *label);
+void	imgui_window_controls(t_rtv *rtv);
 
 #endif
