+++ srcs/loop/loop_sdl.c
@@ -6,7 +6,7 @@
 /*   By: oadhesiv <secondfry+school21@gmail.com>    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/09/05 14:11:10 by oadhesiv          #+#    #+#             */
-/*   Updated: 2021/09/05 15:06:40 by oadhesiv         ###   ########.fr       */
+/*   Updated: 2021/09/07 16:39:33 by oadhesiv         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -15,6 +15,7 @@
 #include "cimgui_impl.h"
 #include <SDL_opengl.h>
 #include "imgui/windows.h"
+#include "imgui/window_controls.h"
 
 static GLuint	create_texture(t_rtv *rtv)
 {
